fix: fix some
