--- tls/cardbox.crt
@@ -1,18 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIC9DCCAdygAwIBAgIJAKSskbR0k7eDMA0GCSqGSIb3DQEBCwUAMBkxFzAVBgNV
-BAMMDmxvY2FsaG9zdDozMDA1MB4XDTIxMDMwNjE2MzEzNloXDTIxMDQwNTE2MzEz
-NlowGTEXMBUGA1UEAwwObG9jYWxob3N0OjMwMDUwggEiMA0GCSqGSIb3DQEBAQUA
-A4IBDwAwggEKAoIBAQDDkRnWC8Hh/K27wH5nV7hHM4z5EtK6BPvyKM7ssy4iKxuf
-ykwxQhPWcIBozjl9MMmKDrkGyQNgylqhBH2EblWILnKRGCdPztPZePvPkY+2qXVF
-TSVbERkq/EkolcoCloaWaTGhzG5U1hBA8/P2wSB5QR0k0uK24X56coMZOb8LKoY/
-mUCEueJupysb+Dvgq6GqNNnAkZ1UCu3nRmal3/aRVMfPTX9gyl48jVxCvELCOKvF
-ME4+3gpDP8Wv8b9HVXTE2VTWpz82jIvvKt9Ru8vw7bU6ZT4+smI690zqXZ+9aNuY
-APhus5+Ct0602OtnUxMRj6kOzyC82Km9PrT1ZZfTAgMBAAGjPzA9MBkGA1UdEQQS
-MBCCDmxvY2FsaG9zdDozMDA1MAsGA1UdDwQEAwIHgDATBgNVHSUEDDAKBggrBgEF
-BQcDATANBgkqhkiG9w0BAQsFAAOCAQEAH/MbXIIab1ReUfet78y+gABlrkLVYkjU
-Flcx8z1oNHW+XDouFmGiYM4NUMkrOXstk0Ykp8bDcG3NlSbCeeLlKS4/WrKDjaMV
-5vLFX8bcQSTanSkisI+OQ5BAn6if0s9o/e4cV3dsBN8/6hnRHXanux2vmHcNfQo9
-yOTmKRMEdRwIrtKsAKTtPir7XEStgglpMdj0a0v9e7QsYiMeKGoMvePSPH5o+6zS
-P+dtgld9QNWIs4ZtXlJeHpbECBGl0vC4JSzlnJFaH6oak1z1CS83DOOHTBxk+CL3
-n+byzhhSiYw4xplyr/sT3xl0X9HPy5dbn+KK9eQ4quEOHZ6Jvm1wog==
------END CERTIFICATE-----
